Simplifying attribute check expressions.
The attribute expressions were repetitive (multiple calls to GetCustomAttributes(false)), and some calls were for the same attributes (ReadOnlyAttribute, DatabaseGeneratedAttribute).

This change reduces the number of GetCustomAttributes() calls to two, and the calls on the array of attributes now use more concise LINQ methods that avoid lambdas in most cases (not avoidable for the dynamic attributes, however).

This is in preparation for adding optimistic concurrency, which checks more attributes.