[DEL] Adding exception for optimistic concurrency failure plus tests.
Throwing an exception is the ideal outcome here; concurrency checks should rarely fail, and if they do, it should be up to the caller to resolve them.

This should also resolve the problem with bulk operations where some items were changed and some were not: *all* items in the statement should be unchanged.